+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+##[1.1.0] - 2017-04-12
+### Added
+- Edit text infos
+- Bonus attribute for texts
+- Reorder texts in a project via Drag & drop
+- Calculate displayedOrder for the UI
+
 ##[1.0.2] - 2017-04-08
 ### Added
 - Error messages 'No Internet connexion' and 'Cannot reach the server'.